requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.2
appdirs==1.4.4
attrs==24.3.0
certifi==2024.12.14
charset-normalizer==3.4.0
et_xmlfile==2.0.0
fake-useragent==2.0.3
frozenlist==1.5.0
idna==3.10
latest-user-agents==0.0.4
multidict==6.1.0
numpy==2.2.0
openpyxl==3.1.5
pandas==2.2.3
propcache==0.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
six==1.17.0
tzdata==2024.2
urllib3==2.2.3
user_agent==0.1.10
yarl==1.18.3
